Adding pahole to enable bpf.
This adds pahole as a dependency in the build context so that when the
kernel is compiled the /sys/kernel/btf directory holds the necessary btf
data.

For more information on this checkout:
https://www.kernel.org/doc/html/latest/bpf/btf.html

To compile a kernel with btf data available you need some config
settings (added in a later commit) and pahole available.

Signed-off-by: Duffie Cooley <dcooley@isovalent.com>