Adding bpf (#43)
* Adding pahole to enable bpf.

This adds pahole as a dependency in the build context so that when the
kernel is compiled the /sys/kernel/btf directory holds the necessary btf
data.

For more information on this checkout:
https://www.kernel.org/doc/html/latest/bpf/btf.html

To compile a kernel with btf data available you need some config
settings (added in a later commit) and pahole available.

Signed-off-by: Duffie Cooley <dcooley@isovalent.com>

* Updated configs to support BPF.

This changes adds some config options for BPF and for BTF so that
tooling that requires BPF can run with these kernels.

Thanks!

Signed-off-by: Duffie Cooley <dcooley@isovalent.com>

---------

Signed-off-by: Duffie Cooley <dcooley@isovalent.com>